+++ NEWS.md
@@ -1,5 +1,7 @@
 # oce 1.7.11
 
+* Remove dependence on `rgeos` and `raster` packages.
+
 * Update `download.topo()` to use new NOAA database.
 
 # oce 1.7.10 (on CRAN)